+++ src/widgets/MenuItem/index.tsx
@@ -2,10 +2,10 @@ import { ReactNode } from 'react';
 import styles from './index.module.css';
 export default function MenuItem({ children, linkUrl } : {children: ReactNode, linkUrl: string}) {
 	return (
-		<a href={linkUrl} className={styles['menu-item']}>
-			<li>
+		<a href={linkUrl} className={styles['menu-item']} aria-label="Ссылка">
+			<ul>
 				{children}
-			</li>
+			</ul>
 		</a>
 	);
 }