+++ src/app/index.tsx
@@ -1,22 +1,15 @@
 import 'slick-carousel/slick/slick-theme.css';
 import 'slick-carousel/slick/slick.css';
 
-import { Articles, Content, LegalServices, NeedConsultation, PracticalArea, QuoteSlider, RequestConsultation } from '../entities';
+import { Articles, Contacts, Content, LegalServices, NeedConsultation, PracticalArea, QuoteSlider, RequestConsultation } from '../entities';
 
-import { useRef } from 'react';
 import { Button, CenterInfoLine, Menu, MenuItem, PhoneNumber, Title } from '../widgets';
-import Input from '../widgets/Input';
 
 function App() {
-	const inputRef = useRef(null);
 	return (
     
 		<>
-			<Input
-				placeholder='Email Address'
-				ref={inputRef}
-				onSend={(textToAction: string) => console.log(textToAction)}
-			/>
+			
     	<CenterInfoLine text={'We have moved into our brand new downtown office at the corner of Broadway and Bellview'}/>
 			<Content> 
 			
@@ -41,6 +34,7 @@ function App() {
 			<NeedConsultation title='Need professional legal advice?' 
 				description='Get a free consultation with our legal experts'
 			/>
+			<Contacts/>
 
 		</>
 	);