+++ src/entities/Contacts/index.tsx
@@ -1,14 +1,15 @@
-import { PhoneNumber, Title } from '../../widgets';
+import { Input, PhoneNumber, Title } from '../../widgets';
+
 import styles from './index.module.css';
 export default function Contacts() {
 	return (
 		<div className={styles['contacts']}>
 			<div className={styles['contacts-left']}>
-				<Title text={'McKinley'} size='h2' color='dark'/>
+				<Title className='title' text={'McKinley'} size='h2' color='light'/>
 				<div className={styles['contacts-info']}>
 					<div className={styles['address']}>2972 Westheimer Road</div>
 					<div className={styles['address']}>Denver, CO 80021</div>
-					<PhoneNumber color='dark' value='(720) 555-0123'/> 
+					<PhoneNumber className={styles['contact-phone-number']} value='(720) 555-0123'/> 
 					<div className={styles['email']}>info.denver@mckinley.com</div>
 				</div>
 				<div className={styles['social-btns']}>
@@ -29,8 +30,11 @@ export default function Contacts() {
 			<div className={styles['contacts-right']}>
 				<Title text={`Get legal news, advice, 
                     and best practices delivered to your inbox.`} 
-				size='h1' color='dark'/>
-				{/**TODO: Input needs to be here  */}
+				size='h1' color='light'/>
+				<Input
+					placeholder='Email Address'
+					onSend={(textToAction: string) => console.log(textToAction)}
+				/>
 			</div>
 		</div>
 	);