+++ src/entities/index.tsx
@@ -1,10 +1,10 @@
 import Articles from './Articles';
+import Contacts from './Contacts';
 import Content from './Content';
 import LegalServices from './LegalServices';
 import NeedConsultation from './NeedConsultation';
 import PracticalArea from './PracticalArea';
 import QuoteSlider from './QuoteSlider';
 import RequestConsultation from './RequestConsultation';
-//TODO: import Contacts from './Contacts';
-export { Articles, Content, LegalServices, NeedConsultation, PracticalArea, QuoteSlider, RequestConsultation };
+export { Articles, Contacts, Content, LegalServices, NeedConsultation, PracticalArea, QuoteSlider, RequestConsultation };