+++ src/widgets/PhoneNumber/index.tsx
@@ -1,8 +1,8 @@
 import cn from 'classnames';
 import styles from './index.module.css';
-export default function PhoneNumber({value, color = 'dark'} : {value: string, color : 'dark' | 'light'}) {
+export default function PhoneNumber({value, color, className} : {value: string, color?: 'dark' | 'light', className?: string}) {
 	return (
-		<div className={cn(styles['phone-number'], {
+		<div className={cn(className, styles['phone-number'], {
 			[styles['dark']]: color == 'dark',
 			[styles['light']]: color == 'light'
 		})}>