+++ src/widgets/index.tsx
@@ -1,14 +1,14 @@
 import Button from './Button';
 import Card from './Card';
 import CenterInfoLine from './CenterInfoLine';
+import Input from './Input';
 import Menu from './Menu';
 import MenuItem from './MenuItem';
 import Paragraph from './Paragraph';
 import PhoneNumber from './PhoneNumber';
 import QuoteSlideItem from './QuoteSlideItem';
 import Title from './Title';
-/**TODO: import Input */
 
 
-export { Button, Card, CenterInfoLine, Menu, MenuItem, Paragraph, PhoneNumber, QuoteSlideItem, Title };
+export { Button, Card, CenterInfoLine, Input, Menu, MenuItem, Paragraph, PhoneNumber, QuoteSlideItem, Title };