+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ##### Chores
 
-- **Contacts:** create structure (1d3aed6c)
+- **Contacts:** create structure (875e0ebe)
 
 ##### New Features
 
+- **Input:** create util Input without css (044c40b8)
 - **NeedConsultation:** add NeedConsultation block (f244b28d)
 - **QuoteSlider:** impl + demo (2009f53e)