+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 ##### New Features
 
 - **Input:**
-  - create Input with css (b3fa22ee)
+  - create Input with css (04a365e1)
   - create util Input without css (38d41f5d)
 - **NeedConsultation:** add NeedConsultation block (f244b28d)
 - **QuoteSlider:** impl + demo (2009f53e)
@@ -17,10 +17,12 @@
 - **Card:**
   - clear warning with a (b8879ecf)
   - clear ts error (a3eeb623)
+  - now text in cards baselined (a3cb679e)
 - **MenuItem:** clear warning (75859083)
 
 ##### Refactors
 
+- renaming style vars (aecac3ea)
 - move to fsd (b90c29c7)
 
 #### 0.0.3 (2024-11-30)